cross prbs; vc: /binobj;

+++ tests/flags_c2.cc
@@ -52,54 +52,54 @@ TEST_CASE("flags test", "[flags]") {
 
     SECTION("--help") {
         const char *argv[] = {"", "--help", "--no-color"};
-        REQUIRE(cli.run(countof(argv), const_cast<char **>(argv)) == 0);
+        REQUIRE(cli.run((int)countof(argv), const_cast<char **>(argv)) == 0);
         REQUIRE(cli.get_for_cli("help").as<bool>());
     }
     SECTION("-h") {
         const char *argv[] = {"", "-h", "--no-color"};
-        REQUIRE(cli.run(countof(argv), const_cast<char **>(argv)) == 0);
+        REQUIRE(cli.run((int) countof(argv), const_cast<char **>(argv)) == 0);
         REQUIRE(cli.get_for_cli("help").as<bool>());
     }
     SECTION("-h-v") {
         const char *argv[] = {"", "-h-v", "--no-color"};
-        REQUIRE(cli.run(countof(argv), const_cast<char **>(argv)) == 0);
+        REQUIRE(cli.run((int) countof(argv), const_cast<char **>(argv)) == 0);
         REQUIRE(cli.get_for_cli("help").as<bool>() == false);
         REQUIRE(cli.get_for_cli("verbose").as<bool>());
     }
 
     SECTION("--version") {
         const char *argv[] = {"", "--version", "--no-color"};
-        REQUIRE(cli.run(countof(argv), const_cast<char **>(argv)) == 0);
+        REQUIRE(cli.run((int) countof(argv), const_cast<char **>(argv)) == 0);
         REQUIRE(cli.get_for_cli("version").as<bool>());
     }
     SECTION("-V") {
         const char *argv[] = {"", "-V", "--no-color"};
-        REQUIRE(cli.run(countof(argv), const_cast<char **>(argv)) == 0);
+        REQUIRE(cli.run((int) countof(argv), const_cast<char **>(argv)) == 0);
         REQUIRE(cli.get_for_cli("version").as<bool>());
     }
     SECTION("--ver") {
         const char *argv[] = {"", "--ver", "--no-color"};
-        REQUIRE(cli.run(countof(argv), const_cast<char **>(argv)) == 0);
+        REQUIRE(cli.run((int) countof(argv), const_cast<char **>(argv)) == 0);
         REQUIRE(cli.get_for_cli("version").as<bool>());
     }
 
 
     SECTION("--verbose") {
         const char *argv[] = {"", "--verbose", "--no-color"};
-        REQUIRE(cli.run(countof(argv), const_cast<char **>(argv)) == 0);
+        REQUIRE(cli.run((int) countof(argv), const_cast<char **>(argv)) == 0);
         REQUIRE(cli.get_for_cli("verbose").as<bool>());
         REQUIRE(cli.get_for_cli("no-color").as<bool>());
     }
     SECTION("-v") {
         const char *argv[] = {"", "-v", "--no-color"};
-        REQUIRE(cli.run(countof(argv), const_cast<char **>(argv)) == 0);
+        REQUIRE(cli.run((int) countof(argv), const_cast<char **>(argv)) == 0);
         REQUIRE(cli.get_for_cli("verbose").as<bool>());
         REQUIRE(cli.get_for_cli("no-color").as<bool>());
         REQUIRE(cli["verbose"].hit_count() == 1);
     }
     SECTION("-vvv") {
         const char *argv[] = {"", "-vvv", "--no-color"};
-        REQUIRE(cli.run(countof(argv), const_cast<char **>(argv)) == 0);
+        REQUIRE(cli.run((int) countof(argv), const_cast<char **>(argv)) == 0);
         REQUIRE(cli.get_for_cli("verbose").as<bool>());
         REQUIRE(cli.get_for_cli("no-color").as<bool>());
         REQUIRE(cli["verbose"].hit_count() == 3);
@@ -108,20 +108,20 @@ TEST_CASE("flags test", "[flags]") {
 
     SECTION("--debug") {
         const char *argv[] = {"", "--debug", "--no-color"};
-        REQUIRE(cli.run(countof(argv), const_cast<char **>(argv)) == 0);
+        REQUIRE(cli.run((int) countof(argv), const_cast<char **>(argv)) == 0);
         REQUIRE(cli.get_for_cli("debug").as<bool>());
         REQUIRE(cli.get_for_cli("no-color").as<bool>());
     }
     SECTION("-D") {
         const char *argv[] = {"", "-D", "--no-color"};
-        REQUIRE(cli.run(countof(argv), const_cast<char **>(argv)) == 0);
+        REQUIRE(cli.run((int) countof(argv), const_cast<char **>(argv)) == 0);
         REQUIRE(cli.get_for_cli("debug").as<bool>());
         REQUIRE(cli.get_for_cli("no-color").as<bool>());
     }
 
     SECTION("~~debug") {
         const char *argv[] = {"", "~~debug", "--no-color"};
-        REQUIRE(cli.run(countof(argv), const_cast<char **>(argv)) == 0);
+        REQUIRE(cli.run((int) countof(argv), const_cast<char **>(argv)) == 0);
         REQUIRE(cli.get_for_cli("debug").as<bool>());
         REQUIRE(cli.get_for_cli("no-color").as<bool>());
         REQUIRE(cli["debug"].hit_special());
@@ -131,7 +131,7 @@ TEST_CASE("flags test", "[flags]") {
 
     SECTION("~~tree") {
         const char *argv[] = {"", "~~tree", "--no-color"};
-        REQUIRE(cli.run(countof(argv), const_cast<char **>(argv)) == 0);
+        REQUIRE(cli.run((int) countof(argv), const_cast<char **>(argv)) == 0);
         REQUIRE(cli.get_for_cli("tree").as<bool>());
         REQUIRE(cli.get_for_cli("no-color").as<bool>());
         REQUIRE(cli["tree"].hit_special());
@@ -143,7 +143,7 @@ TEST_CASE("flags test", "[flags]") {
 
     SECTION("default value of an arg/flag: main --float 2.7") {
         const char *argv[] = {"", "main", "--float", "2.7", "--no-color"};
-        REQUIRE(cli.run(countof(argv), const_cast<char **>(argv)) == 0);
+        REQUIRE(cli.run((int) countof(argv), const_cast<char **>(argv)) == 0);
 
         REQUIRE(cli.get_for_cli("main.int").as<int>() == 10);
         REQUIRE(cli.get_for_cli("main.long").as<long>() == 79129L);
@@ -170,7 +170,7 @@ TEST_CASE("flags test", "[flags]") {
         std::cout << ab << '\n';
 
         const char *argv[] = {"", "main", "--version", "--no-color"};
-        REQUIRE(cli.run(countof(argv), const_cast<char **>(argv)) == 0);
+        REQUIRE(cli.run((int) countof(argv), const_cast<char **>(argv)) == 0);
         REQUIRE(cli.get_for_cli("no-color").as<bool>());
     }
 
@@ -179,8 +179,7 @@ TEST_CASE("flags test", "[flags]") {
     SECTION("required flag: main sub1 --float 2.7") {
         const char *argv[] = {"", "main", "sub1", "--float", "2.7", "--no-color"};
         try {
-            cli.set_no_catch_cmdr_biz_error(true).
-                    run(countof(argv), const_cast<char **>(argv));
+            cli.set_no_catch_cmdr_biz_error(true).run((int) countof(argv), const_cast<char **>(argv));
         } catch (cmdr::exception::cmdr_biz_error const &e) {
             auto c{cmdr::terminal::colors::colorize::create()};
             std::cout << c.bold().s("<<CAPTURED>> ") << e.what() << '\n';
